feat: add support to test kerblam with 'kerblam-examples'
The `kerblam-examples` repo has execution examples for kerblam!.
Since kerblam is an execution-centric tool, not using these examples as
tests feeled like a wasted opportunity.

This commit adds infrastructure to run these tests with `cargo test` by
cloning the 'kerblam-examples' repository, navigating to the example
directory, and executing the bash script `run` in each test directory.

The test expects that the `run` script dies whenever there is an error,
and that it does the checks that we want to do during the test.
It also expects that the files that were in the original example do not
change, as kerblam! should not touch pipes, dockerfiles, etc...