feat(execution): properly manage timestamps of profile files
Before now, when switching profiles, all the timestamps of the moved
files would be updated. This causes `make` to re-run for every profiled
run even though the same run is repeatedly executed.

To avoid this, Kerblam! now makes a file with the name of the last-used
profile in `~/.cache/kerblam/...`. The filename is the hash of the
project path, so that each project gets one cache file.
If kerblam detects that the same profile was used twice in a row, it
skips updating the timestamps.
If a normal run is executed, this cache is deleted and all the input
files are re-touched (as if they were moved) to re-trigger the pipeline.
This should prevent having to `data clean` every time a profile is
switched.