feat: tests should pass much more reliably now
The docker-relying tests used to fail a *lot*. I am not sure of why,
but a hunch is that the parallel execution was messing with something.
Now, with the `serial_test` lib, we run them serially, in just one
thread, and that's it.

Tried them a few times and they have always worked reliably.