fix: fix windows idiotic execution policies
Powershell does not support running files (with `-File`) that do not
end in the `.ps1` extension. This is obviously dumb for interoperability:
you'd have kerblam shell workflows working on linux but not on windows,
and vice-versa.

A workaround is to use `Invoke-Command`, and source the file inside the
"CommandBlock" that Invoke-Command supports. This is fine, but then if
you find out that `Invoke-Command` is *not* a program, *it's a
powershell internal "magic" command*.
So when Kerblam! tries to spawn the `Invoke-Command` executable it
fails.

So, we combine both approaches: we use `powershell -Command` to spawn a
powershell, and pass a command to it. This command is `Invoke-Command`,
which takes care of reading the content of whatever workflow file we are
using, and execute it (in the current shell, so our sub-shell).

Fuck me!