refactor: Removed `reqwest` dependency to speed up compile times
The `reqwest` library is async by defalt, but we used the blocking
module to make blocking calls. However, the blocking module is just a
thin wrapper on the async one. This means that we copile (and use) the
Tokio framework for nothing.

I replaced the `reqwest` dependency with the much more lightweight
`ureq` library.
This gets rid of ~20 dependencies and cuts down compile times,
from about ~70 seconds to ~50 (-28%).